Update assign-preferences.lua
For some food categories, this script saves the mat_state wrong. That is, differently from how the game does it. Dunno if it has any actual effect, apart from Dwarf Therapist showing it in detail. Might actually even be that mat_state should outright be 0, as even drinks don't seem to have 1 for them.

Just start a new embark and look at the preferences of your starting seven with gm-editor or something, I could post screenshots of examples but it should be easy to see for yourself.