Update autowheelbarrow.lua
Changed the amount of days it takes to loop the program as when there are to many wheelbarrows it can take a few days to carry the wheelbarrow over to a stockpile resulting in a infinite loop. 

Changing to 7 days results in even kids reaching their destination with a wheelbarrow before the 7th day.