fix CivBox EditField ignore_keys
The new EditField implementation has a different internal structure.
`ignore_keys` works as part of initialization, but changing it post-init no
longer works. We have to reach "deeper" to change the effective `ignore_keys`
field.

Also, a remove a stray comma. It didn't cause problems, but it was
semantically odd.