Create livingworld.lua
This is my super rudimentary script for checking how many sites of interest there are in the world and printing the result and then increasing the sitecap to allow new sites in post-world generation. Humans build new hamlets, Elves build new Forest Retreats, Goblins build new trenches and Dwarfs build new hillocks.  Ideally the script would increase the sitecap only if the number of sites of interest are +1 -1 away from the sitecap number, but as it is now it works by increasing the sitecap after a set amount of time instead and achieving my goal I hope!