gui/mass-remove.toolbar: compatibility with released default_pos; roll back the fancier positioning
Now that this overlay has been released (51.09-r1), it is important to
go back to a default_pos that matches the released value. This prevents
a "jump" in positioning when trying this code after having used the
mainline version.

If a overlay version change is done in the future, it would probably
make sense to reset the default position to a value that lets the
overlay be moved all the way to the left.

Beyond that, rollback the fancier "apply the same idea in all
directions" code to make the code easier to compare with the what is on
master.

That stuff is probably at least worth thinking about, but not as a part
of simply demonstrating how the "toolbar" library could be used for
toolbar-relative positioning.